Fix YAML spacing
YAML is really cranky about properly spacing things - it took me a
minute to catch what was happening here.

For the `active` and `displayOrder` attributes, the `type` and
`description` fields weren’t correctly aligned with the `name` field.
Notice the very slight spacing differences between the original file
and these changes. If the spacing is incorrect, Jekyll will throw an
error until it’s fixed.