Correct old-schema-template logic in layout/singer.html
If a page didn’t contain a `schema-sections` variable in its
Frontmatter, the correct assignment wasn’t happening. This lead to the
integration schema’s not being displayed even though it was present.

This now checks for `tables` in the integration’s Frontmatter and
reassigns the variable accordingly. This will only be an issue for
non-Singer integration docs.