Re-write the logic for database integration settings fields
Because it was a hot mess.

This commit adds dedicated files for each database type, and each
integration backed by it. It now relies on using a combo of `db-type`
and `name` to locate the correct file and pull the fields. This allows
for flexible field descriptions without trying to shoehorn a bunch of
Liquid into the copy itself.