Fix dates + copy display in version deprecation notice
This adds a new variable to integration version files
(`version.date-last-connection`), which is used in the deprecation
notice to display the `to` and `from` dates for connections.
`date-last-connection` is the date in `MON DD,YYY` format that this
version of the integration could be created in Stitch.

Ex: `Connections created from [date-released] to [date-last-connection]
use this version`.