Support external schema references
Motivation
----------

We need the ability to reference json schema definitions in other files.
This assumes that the external file is called `<file_name>.json` (with the
`.json` extension being required) and is loadable from the same directory
that the rest of the schemas are being loaded from.

This is considerably less extensible than the full spec demands but is all
the we need for now and so can be expanded later.