Fix `api-type` attribute
Some integrations for the Connect API aren’t prefixed with `platform.`,
which was displaying in all descriptions and object examples regardless
of the actual type.

This updates the source form properties to include `platform.` in their
`api-type`, and implements string cleaning in Liquid logic to remove
that prefix when necessary.