Add SSH/SSL to destination guides
While the general destination (ex: Postgres) may support a connection
type, that doesn’t mean it’s supported for specific types (ex: Google
CloudSQL). This is largely dependent on how the provider structures
things, and to keep track of it, it has to be done at the specific
level and not the general one. These variables accomplish that.