Add condition for attribute.name
If a sub attribute doesn’t have a name, just display the description.

This is intended as a temporary solution - we now have the ability to
re-use fields across schemas, thanks to the recursive nature of the new
schema includes. This will be a version 2 improvement, though. For now,
just display descriptions if we aren’t including the entire list of
shared attributes.