Fix display of start_date for Connect (#346)
* Add base docs and setup guide

* Add version file

* Add to Connect

* First pass at schemas

* Fix typo

* Add replication methods and keys

Based on this from the tap’s developer:

```
- lists - full replication
- list_members - date filtered - last modified
- list_segments - full replication
- list segment_members - full replication
- campaigns - full replication
- unsubscribes - full replication
- automations - full replication
- reports_email_activity - date filtered - datetime of activity

Most are full replication because we fetch nested children or filtering
wasn't available
```

* Add more to descriptions

* Add `uses-start-date` to form properties

This controls whether the `start_date` attribute is included in the
source form property.

* Add `uses-start-date` to form property templates

* Update form property attribute logic to use `uses-start-date`

This makes it easier to include the start_date property. It also fixes
an issue where this property wasn’t displaying for form properties that
didn’t have specific attributes.