Shopify - Add attribute
Adds `accepts_marketing_updated_at` field to the customer table.

Context: This is the Sources' PR for this update: https://github.com/singer-io/tap-shopify/pull/69/files
It says that they added this as an `anyOf` field, but the Shopify docs says that this is a date-time field. I put in in the docs as a date-time field too.