TDL-4723: Fix anchor links in changelog
This fixes the issue with anchor links in the changelog. Previously, clicking an anchor link would jump the page to the bottom of the section the anchor was tied to. This was due to some hacky CSS when the changelog was initially created, having to do with how headings display under the sticky topnav.

Entry titles will be actual headings, so the anchor links will work correctly. However, doing this created another issue with the metadata container, causing it to display higher than the entry it's for. *facepalm*

Adding `padding-top: 1px` to .changelog-entry-details fixed that, for whatever reason. If I have to fix this again I may explode.